+++ src/hooks/useIsScroll/Demo/index.tsx
@@ -8,6 +8,7 @@ import { useIsScroll } from 'wwhooks';
 export default () => {
   const ref = useRef();
   const isScrollState = useIsScroll(ref);
+
   return (
     <>
       <p>x : {isScrollState.x.toString()}</p>