+++ src/hooks/useVirtualList/Demo/table.tsx
@@ -1,10 +1,10 @@
 /**
  * debug: true
  */
-import React, { useEffect, useRef } from 'react';
+import React, { useRef } from 'react';
 import useVirtualList from '..';
 import { Button } from 'antd';
-import { useInterval } from '@/';
+import { useIntervalFn } from '../../../';
 
 const mockList = [];
 let key = 0;
@@ -24,9 +24,9 @@ export default () => {
     containerRef,
     30,
   );
-  const { start, clear, looping } = useInterval(() => {
+  const { start, clear, looping } = useIntervalFn(() => {
     const dom = containerRef.current;
-    if (dom.clientHeight <= dom.scrollTop) {
+    if (dom.clientHeight >= dom.scrollTop) {
       clear();
     }
     dom.scrollTop += 100;
@@ -47,7 +47,7 @@ export default () => {
           {cells.map((cell, index) => {
             return (
               <div
-                key={index}
+                key={cell.index}
                 style={{
                   height: 30,
                   display: 'flex',